fix: Duplicate route/model definitions across routers
