+++ backend/routers/knowledge.py
@@ -1,9 +1,10 @@
 """Knowledge Base Router - RAG, Climate Simulation, Seeds"""
-import re
 from typing import Optional
 from fastapi import APIRouter, Request, HTTPException
 from pydantic import BaseModel, Field, validator
 import logging
+
+from backend.schemas import RAGQuery
 
 router = APIRouter()
 logger = logging.getLogger(__name__)
@@ -11,50 +12,6 @@ logger = logging.getLogger(__name__)
 # ---------------------------------------------------------------------------
 # Pydantic models
 # ---------------------------------------------------------------------------
-
-class RAGQuery(BaseModel):
-    query: str = Field(..., min_length=3, max_length=500)
-    top_k: int = Field(default=3, ge=1, le=5)
-
-    @validator("query")
-    def sanitize_and_normalize_query(cls, v):
-        if not v or not isinstance(v, str):
-            raise ValueError("Query must be a non-empty string.")
-
-        v = re.sub(r'<script.*?>.*?</script>', '', v, flags=re.IGNORECASE | re.DOTALL)
-        v = re.sub(r'</?script.*?>', '', v, flags=re.IGNORECASE)
-        v = re.sub(r'on\w+\s*=', '', v, flags=re.IGNORECASE)
-        v = re.sub(r'javascript:', '', v, flags=re.IGNORECASE)
-        v = re.sub(r'data:', '', v, flags=re.IGNORECASE)
-        v = re.sub(r'vbscript:', '', v, flags=re.IGNORECASE)
-        v = re.sub(r'<[^>]*>', '', v)
-        v = re.sub(r'\[([^\]]+)\]\(([^)]+)\)', r'\1 (\2)', v)
-        v = re.sub(r'[*_~`#]', '', v)
-        v = v.strip()
-        v = re.sub(r'\s+', ' ', v)
-
-        forbidden_patterns = [
-            r"ignore\s+(?:all\s+)?previous\s+instructions",
-            r"ignore\s+(?:the\s+)?system\s+prompt",
-            r"override\s+system\s+constraints",
-            r"developer\s+mode",
-            r"bypass\s+safety\s+filter",
-            r"disregard\s+(?:all\s+)?prior\s+instructions",
-            r"act\s+as\s+(?:a\s+)?(?:different|unrestricted|unfiltered)\s+(?:ai|model|assistant)",
-            r"pretend\s+(?:you\s+are|to\s+be)\s+(?:a\s+)?(?:different|unrestricted)",
-            r"jailbreak",
-            r"prompt\s+injection",
-        ]
-        v_lower = v.lower()
-        for pattern in forbidden_patterns:
-            if re.search(pattern, v_lower):
-                raise ValueError("Query contains disallowed phrases or prompt injection attempts.")
-
-        if len(v) < 3:
-            raise ValueError("Query must be at least 3 characters long after sanitization.")
-
-        return v
-
 
 class SimulationRequest(BaseModel):
     """Climate simulation request.
