+++ backend/routers/platform.py
@@ -22,6 +22,8 @@ from reportlab.pdfgen import canvas
 from slowapi import Limiter
 from slowapi.util import get_remote_address
 
+from backend.schemas import AlertTriggerRequest, RAGQuery
+
 router = APIRouter()
 logger = logging.getLogger(__name__)
 limiter = Limiter(key_func=get_remote_address)
@@ -34,11 +36,6 @@ class WhatsAppSubscribeRequest(BaseModel):
     # The authoritative identity is always derived from the verified
     # Firebase ID token — never from client-supplied data.
     user_id: Optional[str] = None
-
-
-class AlertTriggerRequest(BaseModel):
-    alert_type: str = Field(..., pattern=r'^(weather|pest|advisory)$')
-    message: str = Field(..., min_length=1, max_length=500)
 
 
 class ReportRequest(BaseModel):
@@ -62,50 +59,6 @@ class ClientErrorReport(BaseModel):
     source: Optional[str] = Field(default=None, max_length=200)
     stack: Optional[str] = Field(default=None, max_length=2000)
     level: str = Field(default="error", max_length=20)
-
-
-class RAGQuery(BaseModel):
-    query: str = Field(..., min_length=3, max_length=500)
-    top_k: int = Field(default=3, ge=1, le=5)
-
-    @validator("query")
-    def sanitize_and_normalize_query(cls, value):
-        if not value or not isinstance(value, str):
-            raise ValueError("Query must be a non-empty string.")
-
-        value = re.sub(r"<script.*?>.*?</script>", "", value, flags=re.IGNORECASE | re.DOTALL)
-        value = re.sub(r"</?script.*?>", "", value, flags=re.IGNORECASE)
-        value = re.sub(r"on\w+\s*=", "", value, flags=re.IGNORECASE)
-        value = re.sub(r"javascript:", "", value, flags=re.IGNORECASE)
-        value = re.sub(r"data:", "", value, flags=re.IGNORECASE)
-        value = re.sub(r"vbscript:", "", value, flags=re.IGNORECASE)
-        value = re.sub(r"<[^>]*>", "", value)
-        value = re.sub(r"\[([^\]]+)\]\(([^)]+)\)", r"\1 (\2)", value)
-        value = re.sub(r"[*_~`#]", "", value)
-        value = re.sub(r"\s+", " ", value.strip())
-
-        forbidden_patterns = [
-            r"ignore\s+(?:all\s+)?previous\s+instructions",
-            r"ignore\s+(?:the\s+)?system\s+prompt",
-            r"override\s+system\s+constraints",
-            r"developer\s+mode",
-            r"bypass\s+safety\s+filter",
-            r"disregard\s+(?:all\s+)?prior\s+instructions",
-            r"act\s+as\s+(?:a\s+)?(?:different|unrestricted|unfiltered)\s+(?:ai|model|assistant)",
-            r"pretend\s+(?:you\s+are|to\s+be)\s+(?:a\s+)?(?:different|unrestricted)",
-            r"jailbreak",
-            r"prompt\s+injection",
-        ]
-
-        lowered = value.lower()
-        for pattern in forbidden_patterns:
-            if re.search(pattern, lowered):
-                raise ValueError("Query contains disallowed phrases or prompt injection attempts.")
-
-        if len(value) < 3:
-            raise ValueError("Query must be at least 3 characters long after sanitization.")
-
-        return value
 
 
 class GeminiImageRequest(BaseModel):
