+++ backend/schemas/__init__.py
@@ -1,0 +1,4 @@
+"""Shared Pydantic schemas for backend routers."""
+
+from .alerts import AlertTriggerRequest
+from .rag import RAGQuery
